requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.2.0
colorama==0.4.4
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
iniconfig==1.1.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.1
pytest-django==4.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
